Add Ready2Move (r2m) script
The script tags hosts (linked in a specific time range) that can be
"safely" moved from a group to another. A host is safe if there were
no detections in the specified time range.

The tag is randomly generated and is in the form of

   FalconGroupingTags/r2m-********

If no source group is specified (-g) the query is made on all the linked
hosts.

If no time range is specified (--from, --to, --days), the script queries
only on the hosts that were linked in the last 7 days.